v0.9.0.2 (#397)
* docs: add sync script to handle squash merge conflicts

* feat: add new session recovery method

* feat: mark context member as deprecated

* feat: remove go and ts example code

* feat: update get api

* feat: add list api

* feat: update yml

* feat: update yml

* feat: update yml

* feat: update yml

* feat: update yml

* feat: update yml

* feat: update yml

* feat: add screenshot download example

* fix: browser agent undefined attribute

to #70247309

* fix: update go.mod to fix CI/CD build error

Remove toolchain directive from go.mod to resolve the "go mod tidy" error in GitHub Actions CI/CD pipeline.



---------

+++ python/agentbay/browser/eval/run_page_evals.py
@@ -61,7 +61,7 @@ async def run_single_task(
     task_duration_s = round(total_duration_s - browser_setup_s, 2)
 
     llm_metrics = agent.get_metrics()
-    llm_metrics["llm_duration_s"] = round(llm_metrics.get("llm_duration_s", 0.0), 2)
+    llm_metrics["llm_duration_s"] = int(round(llm_metrics.get("llm_duration_s", 0.0), 2))
     if task_duration_s:
         llm_time_percentage = round(
             llm_metrics["llm_duration_s"] / task_duration_s * 100, 1
